test(kona-sp1): share super-aggregation fixture
Expose the canonical two-chain input behind a narrow test-only feature so client validation and guest digest tests cannot drift.
