+++ rust/kona/sp1/crates/client/Cargo.toml
@@ -45,6 +45,9 @@ kzg-rs.workspace = true
 cfg-if.workspace = true
 
 [features]
+# Exposes the canonical super-aggregation fixture to downstream guest tests. Kept narrowly named
+# to avoid Zepter propagating unrelated dependency test features.
+aggregation-test-utils = []
 # Exposes witness-tampering helpers (e.g. `PreimageStore::overwrite_preimage`) used by host-side
 # test tooling. Never enabled by the guest program. Deliberately not named `test-utils`: that would
 # trigger Zepter's propagate-feature check and drag the dependencies' `test-utils` (incl. RPC deps
