+++ rust/kona/sp1/crates/client/src/lib.rs
@@ -11,6 +11,9 @@ pub mod range;
 
 pub mod super_root;
 
+#[cfg(any(test, feature = "aggregation-test-utils"))]
+pub mod test_utils;
+
 pub mod types;
 
 extern crate alloc;
