+++ rust/kona/sp1/programs/super-aggregation/Cargo.toml
@@ -15,5 +15,8 @@ sha2.workspace = true
 sp1-lib.workspace = true
 sp1-zkvm.workspace = true
 
+[dev-dependencies]
+kona-sp1-client-utils = { workspace = true, features = ["aggregation-test-utils"] }
+
 [lints]
 workspace = true
